Aplicacion de middleares y separacion de menu por tipo de usuario mas registro de horario medico

+++ routes/web.php
@@ -22,18 +22,26 @@ Auth::routes();
 Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
 
 
-//Rutas para especialidades
-Route::get('/especialidades',[App\Http\Controllers\EspecialidadController::class, 'index']);
-Route::get('/especialidades/crear',[App\Http\Controllers\EspecialidadController::class, 'crear']); //muestra el formulario de registro
-Route::get('/especialidades/{especialidad}/editar',[App\Http\Controllers\EspecialidadController::class, 'editar']);//nos dirige a la vista de editar
-Route::post('/especialidades',[App\Http\Controllers\EspecialidadController::class, 'guardar']); //envia el fromulario de registro
-Route::put('/especialidades/{especialidad}',[App\Http\Controllers\EspecialidadController::class, 'actualizar']); //actualiza una especialidad
-Route::delete('/especialidades/{especialidad}',[App\Http\Controllers\EspecialidadController::class, 'eliminar']); //elimina una especialidad
+Route::middleware(['auth','admin'])->group(function () {
+    Route::get('/especialidades',[App\Http\Controllers\Admin\EspecialidadController::class, 'index']);
+    Route::get('/especialidades/crear',[App\Http\Controllers\Admin\EspecialidadController::class, 'crear']); //muestra el formulario de registro
+    Route::get('/especialidades/{especialidad}/editar',[App\Http\Controllers\Admin\EspecialidadController::class, 'editar']);//nos dirige a la vista de editar
+    Route::post('/especialidades',[App\Http\Controllers\Admin\EspecialidadController::class, 'guardar']); //envia el fromulario de registro
+    Route::put('/especialidades/{especialidad}',[App\Http\Controllers\Admin\EspecialidadController::class, 'actualizar']); //actualiza una especialidad
+    Route::delete('/especialidades/{especialidad}',[App\Http\Controllers\Admin\EspecialidadController::class, 'eliminar']); //elimina una especialidad
 
-//Rutas para Médicos
-Route::get('/medicos',[App\Http\Controllers\MedicoController::class, 'index']);
-Route::get('/medicos/crear',[App\Http\Controllers\MedicoController::class, 'crear']); //muestra el formulario de registro
-Route::get('/medicos/{medico}/editar',[App\Http\Controllers\MedicoController::class, 'editar']);//nos dirige a la vista de editar
-Route::post('/medicos',[App\Http\Controllers\MedicoController::class, 'guardar']); //envia el fromulario de registro
-Route::put('/medicos/{medico}',[App\Http\Controllers\MedicoController::class, 'actualizar']); //actualiza un medico
-Route::delete('/medicos/{medico}',[App\Http\Controllers\MedicoController::class, 'eliminar']); //elimina un medico
+    //Rutas para Médicos
+    Route::get('/medicos',[App\Http\Controllers\Admin\MedicoController::class, 'index']);
+    Route::get('/medicos/crear',[App\Http\Controllers\Admin\MedicoController::class, 'crear']); //muestra el formulario de registro
+    Route::get('/medicos/{medico}/editar',[App\Http\Controllers\Admin\MedicoController::class, 'editar']);//nos dirige a la vista de editar
+    Route::post('/medicos',[App\Http\Controllers\Admin\MedicoController::class, 'guardar']); //envia el fromulario de registro
+    Route::put('/medicos/{medico}',[App\Http\Controllers\Admin\MedicoController::class, 'actualizar']); //actualiza un medico
+    Route::delete('/medicos/{medico}',[App\Http\Controllers\Admin\MedicoController::class, 'eliminar']); //elimina un medico
+});
+
+
+Route::middleware(['auth','medico'])->group(function () {
+   
+    Route::get('/horario',[App\Http\Controllers\Medico\HorarioController::class, 'editar']);
+    Route::post('/horario',[App\Http\Controllers\Medico\HorarioController::class, 'guardar']);
+});
